Predict/requirements.txt
azure-functions

numpy
h5py
scikit-learn
keras
tensorflow
pandas
